float

+++ .vscode/DevisonOperator.java
@@ -1,9 +1,9 @@
 public class DevisonOperator {
     public static void main(String[] args) {
-        int a = 5;
-        int b = 3;
-        int c = a / b;
-        int s = a % b;
+        float a = 5;
+        float b = 3;
+        float c = a / b;
+        float s = a % b;
 
         System.out.println(c);
         System.out.println(s);
